pyproject.toml
[tool.poetry]
name = "mai-db-cp"
version = "0.1.0"
description = "Course project MAI DB"
authors = ["Kirill Efimenko <vivichv9@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
bcrypt = "^4.2.1"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
